requirements.txt
absl-py==1.3.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
einops==0.4.1
fvcore==0.1.5
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.10
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
pandas==1.1.5
Pillow==8.4.0
pkg_resources==0.0.0
portalocker==2.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.18.0
six==1.16.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.13
Werkzeug==2.0.3
yacs==0.1.8
zipp==3.6.0